+++ packages/moonstone/CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Fixed
 
-- `moonstone/Panels.Panel` to restore focus properly when panel has `moonstone/Scroller` with `focusableScrollbar`
+- `moonstone/Panels.Panel` to restore focus properly when it has `moonstone/Scroller` with `focusableScrollbar`
 
 ## [3.2.6] - 2020-03-26